requirements.txt
streamlit
pydantic
graphviz
python-pptx
openai
google-genai
python-docx
pypdf
pillow
